+++ bee-storage/bee-storage/src/backend.rs
@@ -4,8 +4,6 @@
 use async_trait::async_trait;
 use serde::de::DeserializeOwned;
 
-use std::error::Error;
-
 /// Trait to be implemented on a storage backend.
 /// Determines how to start and shutdown the backend.
 #[async_trait]
@@ -18,11 +16,11 @@ pub trait StorageBackend: Send + Sized + Sync + 'static {
     type Error: std::error::Error + Send;
 
     /// Initializes and starts the backend.
-    async fn start(config: Self::Config) -> Result<Self, Box<dyn Error>>;
+    async fn start(config: Self::Config) -> Result<Self, Self::Error>;
 
     /// Shutdowns the backend.
-    async fn shutdown(self) -> Result<(), Box<dyn Error>>;
+    async fn shutdown(self) -> Result<(), Self::Error>;
 
     /// Size of the database in bytes.
-    async fn size(&self) -> Result<usize, Box<dyn Error>>;
+    async fn size(&self) -> Result<usize, Self::Error>;
 }