+++ bee-storage/bee-storage/CHANGELOG.md
@@ -21,8 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.2.0-alpha - 2021-01-11
 
+### Added
+
 - `AsStream::Stream: Send + Sync + Unpin` bounds;
 - `StorageBackend::size()` method;
+
+### Changed
+
+- `StorageBackend` methods return `Result<_, Self::Error>` instead of `Result<_, Box<dyn std::error::Error>>`;
 
 ## 0.1.0-alpha - 2021-01-06